Replace MakeEpochXid macro.
This macro only works if the epoch is passed as a uint64, but the
normal type for an epoch is uint32, which led to code contortions
in a number of places.  Replace this macro with an inline function,
called FullTransactionIdFromEpochAndXid for Thomas Munro's recent
work on the PostgreSQL master branch.

Rip out most of the places where epochs are stored as uint64 rather
than uint32, but leave FetchTransInfoFromUndo alone for the time
being, since it needs more extensive surgery.

+++ src/backend/access/zheap/zundo.c
@@ -545,7 +545,7 @@ zbuffer_exec_pending_rollback(Relation rel, Buffer buf, BlockNumber *tpd_blkno)
 {
 	int			slot_no;
 	int			total_trans_slots = 0;
-	uint64		epoch;
+	uint32		epoch;
 	TransactionId xid;
 	UndoRecPtr	urec_ptr;
 	TransInfo  *trans_slots = NULL;
